moved all code into reconcile.statuspage
