+++ reconcile/statuspage/integration.py
@@ -4,8 +4,8 @@ import sys
 from reconcile import queries
 from reconcile.utils.semver_helper import make_semver
 from reconcile.utils.state import State
-from reconcile.utils.statuspage import atlassian
-from reconcile.utils.statuspage.models import (
+from reconcile.statuspage import atlassian
+from reconcile.statuspage.models import (
     StatusPage,
     register_provider,
 )
